Fix wrong error handling
If one of the required tools is not installed the script reports this
error but fails to terminate. This is because the round brackets around
the "body" of the "if condition" spawn a subshell which runs the echo
command and then exists with error code 3. However, this does not affect
the exit behaviour of the "main" shell which happily continues to run
the check and even exits with 0.

Remove the brackets so the exit command runs on the main shell.